Fixed verify notes
