+++ src/main/kotlin/com/angorasix/notifications/infrastructure/constants/ApiConstants.kt
@@ -6,11 +6,16 @@ package com.angorasix.notifications.infrastructure.constants
  *
  * @author rozagerardo
  */
-enum class ApiConstants(val value: String) {
-    IDS_QUERY_PARAM("ids"),
-    DISMISSED_QUERY_PARAM("dismissed"),
-    PAGE_QUERY_PARAM("page"),
-    PAGESIZE_QUERY_PARAM("pageSize"),
-    SORT_QUERY_PARAM("sort"),
-    EXTRA_SKIP_QUERY_PARAM("extraSkip"),
+class ApiConstants private constructor() {
+    companion object {
+        const val IDS_QUERY_PARAM = "ids"
+        const val DISMISSED_QUERY_PARAM = "dismissed"
+        const val PAGE_QUERY_PARAM = "page"
+        const val PAGE_DEFAULT_VALUE = 0
+        const val PAGESIZE_QUERY_PARAM = "pageSize"
+        const val PAGESIZE_DEFAULT_VALUE = 20
+        const val SORT_QUERY_PARAM = "sort"
+        const val EXTRA_SKIP_QUERY_PARAM = "extraSkip"
+        const val EXTRA_SKIP_DEFAULT_VALUE = 0
+    }
 }
